Discriminator

+++ dataset.py
@@ -18,7 +18,7 @@ def read_pgm(pgmf):
 		for y in range(width):
 			row.append(ord(pgmf.read(1)))
 		raster.append(row)
-	return torch.tensor(raster, dtype=torch.uint8).unsqueeze(0)
+	return torch.tensor(raster, dtype=torch.uint8).unsqueeze(0).float()
 
 class Dataset(torch.utils.data.Dataset):
